refactor: move InputValidatorTest to its own file and add a test to check if winning numbers input has any duplicated number
