+++ src/test/kotlin/lotto/ApplicationTest.kt
@@ -4,7 +4,6 @@ import camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersI
 import camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest
 import camp.nextstep.edu.missionutils.test.NsTest
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 
@@ -57,58 +56,5 @@ class ApplicationTest : NsTest() {
 
     companion object {
         private const val ERROR_MESSAGE: String = "[ERROR]"
-    }
-}
-
-class InputValidatorTest : NsTest() {
-    @Test
-    fun `assert that input isn't empty`() {
-        val validator = InputValidator()
-
-        assertSimpleTest {
-            assertThrows<IllegalArgumentException> { validator.validateBudget("") }
-        }
-    }
-
-    @Test
-    fun `assert that input isn't less than 1`() {
-        val validator = InputValidator()
-
-        assertSimpleTest {
-            assertThrows<IllegalArgumentException> { validator.validateBudget("0") }
-            assertThrows<IllegalArgumentException> { validator.validateBudget("-1") }
-        }
-    }
-
-    @Test
-    fun `assert that input is divisible by 1000`() {
-        val validator = InputValidator()
-
-        assertSimpleTest {
-            assertThrows<IllegalArgumentException> { validator.validateBudget("5") }
-            assertThrows<IllegalArgumentException> { validator.validateBudget("3.3") }
-        }
-    }
-
-    @Test
-    fun `assert that winning numbers are not empty`() {
-        val validator = InputValidator()
-        assertThrows<IllegalArgumentException> { validator.validateWinningNumber("") }
-    }
-
-    @Test
-    fun `validateWinningNumber throws exception when input has less than 6 numbers`() {
-        val validator = InputValidator()
-        assertThrows<IllegalArgumentException> { validator.validateWinningNumber("1,2,3,4,5") }
-    }
-
-    @Test
-    fun `validateWinningNumber throws exception when input isn't between 1 and 45`() {
-        val validator = InputValidator()
-        assertThrows<IllegalArgumentException> { validator.validateWinningNumber("1,2,3,4,47,5")}
-    }
-
-    override fun runMain() {
-        main()
     }
 }
